+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/my/MyChannelsDelegatee.scala
@@ -5,7 +5,8 @@ import android.content.Context
 import android.support.v4.app.FragmentActivity
 import android.support.v7.app.AlertDialog
 import android.support.v7.widget.LinearLayoutManager
-import x7c1.linen.database.{ChannelSubscriber, LinenOpenHelper}
+import x7c1.linen.database.LinenOpenHelper
+import x7c1.linen.database.struct.ChannelSubscriber
 import x7c1.linen.glue.activity.ActivityControl
 import x7c1.linen.glue.activity.ActivityLabel.SettingMyChannelSources
 import x7c1.linen.glue.res.layout.{SettingMyChannelCreate, SettingMyChannelsLayout}
