+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/AttachSourceDialog.scala
@@ -10,7 +10,8 @@ import android.support.v7.widget.RecyclerView.Adapter
 import android.view.ViewGroup
 import android.widget.CompoundButton.OnCheckedChangeListener
 import android.widget.{Button, CompoundButton}
-import x7c1.linen.database.{ChannelSourceMapKey, ChannelSourceMapParts, LinenOpenHelper}
+import x7c1.linen.database.struct.{ChannelSourceMapKey, ChannelSourceMapParts}
+import x7c1.linen.database.LinenOpenHelper
 import x7c1.linen.glue.res.layout.{SettingSourceAttach, SettingSourceAttachRow, SettingSourceAttachRowItem}
 import x7c1.linen.modern.init.settings.preset.AttachSourceDialog.Arguments
 import x7c1.linen.repository.channel.my.{ChannelToAttach, ChannelsToAttachAccessor}
