+++ linen-modern/src/main/scala/x7c1/linen/modern/init/updater/LoadedSource.scala
@@ -1,6 +1,7 @@
 package x7c1.linen.modern.init.updater
 
-import x7c1.linen.database.{EntryParts, SourceRecord, Updatable}
+import x7c1.linen.database.struct.{SourceRecord, EntryParts}
+import x7c1.linen.database.Updatable
 import x7c1.wheat.macros.database.TypedFields
