+++ linen-modern/src/main/scala/x7c1/linen/modern/init/updater/SourceInspector.scala
@@ -5,7 +5,8 @@ import java.net.{HttpURLConnection, URL}
 
 import com.google.code.rome.android.repackaged.com.sun.syndication.feed.synd.{SyndEntry, SyndFeed}
 import com.google.code.rome.android.repackaged.com.sun.syndication.io.SyndFeedInput
-import x7c1.linen.database.{EntryParts, SourceRecord, LinenOpenHelper}
+import x7c1.linen.database.struct.{SourceRecord, EntryParts}
+import x7c1.linen.database.LinenOpenHelper
 import x7c1.linen.repository.date.Date
 import x7c1.linen.repository.entry.EntryUrl
 import x7c1.wheat.modern.callback.CallbackTask
