+++ linen-repository/src/main/scala/x7c1/linen/database/mixin/TaggedAccountRecord.scala
@@ -1,6 +1,7 @@
 package x7c1.linen.database.mixin
 
-import x7c1.linen.database.{Query, AccountTagLabel}
+import x7c1.linen.database.Query
+import x7c1.linen.database.struct.AccountTagLabel
 import x7c1.wheat.macros.database.TypedFields
 
 trait TaggedAccountRecord extends TypedFields {
