+++ linen-repository/src/main/scala/x7c1/linen/database/struct/AccountRecord.scala
@@ -1,5 +1,6 @@
-package x7c1.linen.database
+package x7c1.linen.database.struct
 
+import x7c1.linen.database.Insertable
 import x7c1.linen.repository.date.Date
 import x7c1.wheat.macros.database.TypedFields
