+++ linen-repository/src/main/scala/x7c1/linen/database/struct/EntryRecord.scala
@@ -1,6 +1,7 @@
-package x7c1.linen.database
+package x7c1.linen.database.struct
 
 import android.content.ContentValues
+import x7c1.linen.database.Insertable
 import x7c1.linen.repository.date.Date
 import x7c1.linen.repository.entry.EntryUrl
 import x7c1.wheat.macros.database.TypedFields
