+++ linen-repository/src/main/scala/x7c1/linen/database/struct/SourceRecord.scala
@@ -1,7 +1,8 @@
-package x7c1.linen.database
+package x7c1.linen.database.struct
 
 import android.content.ContentValues
 import android.database.Cursor
+import x7c1.linen.database.{Insertable, SingleWhere}
 import x7c1.linen.repository.date.Date
 import x7c1.wheat.macros.database.{TypedCursor, TypedFields}
