+++ linen-repository/src/main/scala/x7c1/linen/database/struct/retrieved_source_marks.scala
@@ -1,6 +1,7 @@
-package x7c1.linen.database
+package x7c1.linen.database.struct
 
 import android.database.Cursor
+import x7c1.linen.database.SingleWhere
 import x7c1.linen.repository.date.Date
 import x7c1.wheat.macros.database.{TypedCursor, TypedFields}
