+++ linen-repository/src/main/scala/x7c1/linen/repository/channel/my/ChannelsToAttachAccessor.scala
@@ -2,7 +2,8 @@ package x7c1.linen.repository.channel.my
 
 import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
-import x7c1.linen.database.{ChannelRecord, Query}
+import x7c1.linen.database.Query
+import x7c1.linen.database.struct.ChannelRecord
 import x7c1.wheat.macros.database.TypedCursor
 import x7c1.wheat.modern.sequence.Sequence
