+++ linen-repository/src/main/scala/x7c1/linen/repository/channel/preset/PresetChannel.scala
@@ -1,7 +1,8 @@
 package x7c1.linen.repository.channel.preset
 
 import android.database.Cursor
-import x7c1.linen.database.{ChannelRecord, SingleWhere}
+import x7c1.linen.database.SingleWhere
+import x7c1.linen.database.struct.ChannelRecord
 import x7c1.linen.repository.account.PresetAccount
 import x7c1.linen.repository.source.setting.Channel
 import x7c1.wheat.macros.database.TypedCursor
