+++ linen-repository/src/main/scala/x7c1/linen/repository/entry/unread/EntryAccessor.scala
@@ -3,7 +3,8 @@ package x7c1.linen.repository.entry.unread
 import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
 import android.support.v7.widget.RecyclerView.ViewHolder
-import x7c1.linen.database.{EntryRecord, Query}
+import x7c1.linen.database.Query
+import x7c1.linen.database.struct.EntryRecord
 import x7c1.linen.repository.unread.{EntryKind, SourceKind, UnreadRowKind}
 import x7c1.wheat.macros.database.TypedCursor
 import x7c1.wheat.macros.logger.Log
