+++ linen-repository/src/main/scala/x7c1/linen/repository/source/unread/SourceTitle.scala
@@ -2,8 +2,9 @@ package x7c1.linen.repository.source.unread
 
 import android.content.ContentValues
 import android.database.Cursor
-import x7c1.linen.database.SourceRecord.table
-import x7c1.linen.database.{SingleWhere, SourceRecord, Updatable}
+import x7c1.linen.database.struct.SourceRecord
+import SourceRecord.table
+import x7c1.linen.database.{SingleWhere, Updatable}
 import x7c1.wheat.macros.database.{TypedCursor, TypedFields}
 
 case class SourceTitle(
