+++ linen-repository/src/test/scala/x7c1/linen/database/EntryPartsTest.scala
@@ -6,6 +6,7 @@ import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
+import x7c1.linen.database.struct.{retrieved_source_marks, EntryParts}
 import x7c1.linen.repository.date.Date
 import x7c1.linen.repository.entry.EntryUrl
 import x7c1.linen.repository.source.unread.UnreadSourceFixture
