+++ linen-repository/src/test/scala/x7c1/linen/repository/account/preset/PresetAccountSetupTest.scala
@@ -6,7 +6,8 @@ import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
-import x7c1.linen.database.{LinenOpenHelper, account_tag_map, account_tags}
+import x7c1.linen.database.struct.{account_tags, account_tag_map}
+import x7c1.linen.database.LinenOpenHelper
 import x7c1.linen.repository.account.setup.PresetAccountSetup
 import x7c1.linen.repository.source.setting.SampleFactory
