+++ linen-repository/src/test/scala/x7c1/linen/repository/channel/preset/PresetChannelsAccessorTest.scala
@@ -6,7 +6,8 @@ import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
-import x7c1.linen.database.{ChannelSubscriber, LinenOpenHelper}
+import x7c1.linen.database.LinenOpenHelper
+import x7c1.linen.database.struct.ChannelSubscriber
 import x7c1.linen.repository.account.setup.ClientAccountSetup
 import x7c1.linen.repository.preset.PresetFactory
