+++ linen-repository/src/test/scala/x7c1/linen/repository/source/setting/SettingSourceAccessorTest.scala
@@ -7,7 +7,8 @@ import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
-import x7c1.linen.database.{AccountParts, ChannelParts, LinenOpenHelper, QueryExplainer}
+import x7c1.linen.database.struct.{ChannelParts, AccountParts}
+import x7c1.linen.database.{LinenOpenHelper, QueryExplainer}
 import x7c1.linen.repository.account.dev.Account
 import x7c1.linen.repository.date.Date
